refactor(backoffice-v2): replaced the old form component with the new shadcn one
